Added automatic channel switching in receiver mode.
In receiver mode, added automatic switching to channel 0 if valid DCC is not detected on the current RF channel within 10 seconds after startup (approximately). This allows the user to be certain of the receiver's channel number if unknown by waiting 10 or so seconds before turning on the transmitter and setting to channel 0. The channel change is NOT permanent unless the user resets CV255 to a new channel number in OPS mode at the receiver's DCC address (default=9001).