Update AirMiniSketchTransmitter_Nmra.ino
Rationalized the # of preamble bits. If preamble_bits (=CV242) is >= to 30, then we want Airwire compatibility, so we add a 1/2 "1" 1/2 "0" CUTOUT that the Airwire products seem to "like."

Setting CV242 to < 16 causes the transmitter to use the # of preamble bits from the input message. Otherwise, the # of transmitter preamble bits has the value of CV242.